Incorporate preload/run_calculator.ts into preload/script.ts
